+++ django_elastipymemcache/client.py
