+++ django_elastipymemcache/cluster_utils.py
